Moves playlists and tracks components to under routes folder
